Updates to pre pipeline
Added an experimental function to collect the embedded images in tweets. needs to be updated to check for other types of media also. The pipeline needs to be updated to add this to the other scripts. Fixed the -1 issue in the scripts .js loop. WE NEED DYNAMIC DIRECTORIES right now I cannot run things since they are fixed directories for your file system :)