Updates to features and questions
I have added new experimental features to make the platform more realistic. I have also commented on many lines so we can discuss tonight what it is they do. I think I get alot of it but some seems to be inefficient since we have to write out to files constantly... is there a faster way?