AttributeError Exception
The name of the exception that we have been catching is known as AttributeError, so it has been added. Furthermore, previously we checked if imageLink was none, we attempt to try twitter:image:src. However, when checking for the name of the exception, it was discovered that the exception occurred when trying to call the .get() function, so the exception may have occurred even before attempting to check if imageLink was None. This should cover it