CardInfo Title&Description HTML Fix
Yesterday I discovered that some article previews may have html script within it. Though this issue is rare, it's still possible as I discovered yesterday, so I applied Beautiful Soup on the content of the title and description, and if it contains any HTML, it will parse that out and obtain solely the text.